refactor(thread): move branch-name generator next to create tool
