+++ apps/mesh/src/tools/thread/branch-name.ts
@@ -1,0 +1,120 @@
+/**
+ * Branch-name generator used as a fallback when VM_START is invoked without
+ * an explicit branch. The word lists were lifted out of daemon.ts so the
+ * orchestrator — not the sandbox — decides the branch name and can persist
+ * it to vmMap before the daemon ever sees it.
+ *
+ * Format: `deco/<adjective>-<noun>`
+ */
+
+const ADJECTIVES: readonly string[] = [
+  "amber",
+  "bold",
+  "bright",
+  "calm",
+  "crimson",
+  "coral",
+  "daring",
+  "deep",
+  "dusty",
+  "eager",
+  "faint",
+  "fierce",
+  "frozen",
+  "gentle",
+  "golden",
+  "grand",
+  "green",
+  "hollow",
+  "iron",
+  "ivory",
+  "keen",
+  "lasting",
+  "lunar",
+  "mellow",
+  "misty",
+  "noble",
+  "olive",
+  "pale",
+  "prime",
+  "quiet",
+  "rapid",
+  "rustic",
+  "serene",
+  "sharp",
+  "silver",
+  "sleek",
+  "solar",
+  "stark",
+  "still",
+  "swift",
+  "tawny",
+  "tender",
+  "thin",
+  "true",
+  "vast",
+  "velvet",
+  "warm",
+  "wild",
+  "young",
+  "zen",
+];
+
+const NOUNS: readonly string[] = [
+  "anchor",
+  "birch",
+  "brook",
+  "cedar",
+  "cliff",
+  "cove",
+  "crane",
+  "dune",
+  "echo",
+  "ember",
+  "falcon",
+  "fern",
+  "flint",
+  "forge",
+  "frost",
+  "glade",
+  "grove",
+  "harbor",
+  "hawk",
+  "iris",
+  "jade",
+  "lark",
+  "maple",
+  "marsh",
+  "mesa",
+  "opal",
+  "orbit",
+  "peak",
+  "pine",
+  "plume",
+  "quartz",
+  "rapids",
+  "reef",
+  "ridge",
+  "river",
+  "sage",
+  "shore",
+  "slate",
+  "spruce",
+  "stone",
+  "summit",
+  "thorn",
+  "tide",
+  "trail",
+  "vale",
+  "wren",
+  "aspen",
+  "delta",
+  "crest",
+  "spark",
+];
+
+export function generateBranchName(): string {
+  const adj = ADJECTIVES[Math.floor(Math.random() * ADJECTIVES.length)];
+  const noun = NOUNS[Math.floor(Math.random() * NOUNS.length)];
+  return `deco/${adj}-${noun}`;
+}
